:truck: move the widget into its own module
